Recording videos for ppo and player, some refactoring.
